requirements.txt
geopandas
mapclassify
statsmodels
scipy
itertools
seaborn